refactor(json-crdt-patch): 💡 consolidate operations in one file
